pyproject.toml
[project]
name = "flux-tool"
version = "1.5.0"
requires-python = ">=3.13"
description = "Package used to study the output histograms from PPFX"
authors = [{name = "Anthony Wood", email = "apwood@uh.edu"}]
readme = "README.md"
dependencies = [
    "hist>=2.8.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.0",
    "mplhep>=0.3.55",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "rich>=13.9.4",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
    "uproot>=5.5.1",
]

[project.scripts]
flux_tool = "flux_tool:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
